feat: add solutions to lc problem: No.0480
No.0480.Sliding Window Median
