+++ solution/0400-0499/0480.Sliding Window Median/Solution2.java
@@ -1,0 +1,43 @@
+class Solution {
+    public double[] medianSlidingWindow(int[] nums, int k) {
+        TreeMap<Integer, Integer> l = new TreeMap<>();
+        TreeMap<Integer, Integer> r = new TreeMap<>();
+        int n = nums.length;
+        double[] ans = new double[n - k + 1];
+        int lSize = 0, rSize = 0;
+        for (int i = 0; i < n; ++i) {
+            r.merge(nums[i], 1, Integer::sum);
+            int x = r.firstKey();
+            if (r.merge(x, -1, Integer::sum) == 0) {
+                r.remove(x);
+            }
+            l.merge(x, 1, Integer::sum);
+            ++lSize;
+            while (lSize - rSize > 1) {
+                x = l.lastKey();
+                if (l.merge(x, -1, Integer::sum) == 0) {
+                    l.remove(x);
+                }
+                r.merge(x, 1, Integer::sum);
+                --lSize;
+                ++rSize;
+            }
+            int j = i - k + 1;
+            if (j >= 0) {
+                ans[j] = k % 2 == 1 ? l.lastKey() : ((double) l.lastKey() + r.firstKey()) / 2;
+                if (l.containsKey(nums[j])) {
+                    if (l.merge(nums[j], -1, Integer::sum) == 0) {
+                        l.remove(nums[j]);
+                    }
+                    --lSize;
+                } else {
+                    if (r.merge(nums[j], -1, Integer::sum) == 0) {
+                        r.remove(nums[j]);
+                    }
+                    --rSize;
+                }
+            }
+        }
+        return ans;
+    }
+}
