+++ solution/0400-0499/0480.Sliding Window Median/Solution2.py
@@ -1,0 +1,21 @@
+from sortedcontainers import SortedList
+
+
+class Solution:
+    def medianSlidingWindow(self, nums: List[int], k: int) -> List[float]:
+        l = SortedList()
+        r = SortedList()
+        ans = []
+        for i, x in enumerate(nums):
+            r.add(x)
+            l.add(r.pop(0))
+            while len(l) - len(r) > 1:
+                r.add(l.pop())
+            j = i - k + 1
+            if j >= 0:
+                ans.append(l[-1] if k & 1 else (l[-1] + r[0]) / 2)
+                if nums[j] in l:
+                    l.remove(nums[j])
+                else:
+                    r.remove(nums[j])
+        return ans
