+++ solution/0400-0499/0480.Sliding Window Median/Solution2.ts
@@ -1,629 +1,3 @@
----
-comments: true
-difficulty: Hard
-edit_url: https://github.com/doocs/leetcode/edit/main/solution/0400-0499/0480.Sliding%20Window%20Median/README_EN.md
-tags:
-    - Array
-    - Hash Table
-    - Sliding Window
-    - Heap (Priority Queue)
----
-
-<!-- problem:start -->
-
-# [480. Sliding Window Median](https://leetcode.com/problems/sliding-window-median)
-
-[中文文档](/solution/0400-0499/0480.Sliding%20Window%20Median/README.md)
-
-## Description
-
-<!-- description:start -->
-
-<p>The <strong>median</strong> is the middle value in an ordered integer list. If the size of the list is even, there is no middle value. So the median is the mean of the two middle values.</p>
-
-<ul>
-	<li>For examples, if <code>arr = [2,<u>3</u>,4]</code>, the median is <code>3</code>.</li>
-	<li>For examples, if <code>arr = [1,<u>2,3</u>,4]</code>, the median is <code>(2 + 3) / 2 = 2.5</code>.</li>
-</ul>
-
-<p>You are given an integer array <code>nums</code> and an integer <code>k</code>. There is a sliding window of size <code>k</code> which is moving from the very left of the array to the very right. You can only see the <code>k</code> numbers in the window. Each time the sliding window moves right by one position.</p>
-
-<p>Return <em>the median array for each window in the original array</em>. Answers within <code>10<sup>-5</sup></code> of the actual value will be accepted.</p>
-
-<p>&nbsp;</p>
-<p><strong class="example">Example 1:</strong></p>
-
-<pre>
-<strong>Input:</strong> nums = [1,3,-1,-3,5,3,6,7], k = 3
-<strong>Output:</strong> [1.00000,-1.00000,-1.00000,3.00000,5.00000,6.00000]
-<strong>Explanation:</strong>
-Window position                Median
----------------                -----
-[<strong>1  3  -1</strong>] -3  5  3  6  7        1
- 1 [<strong>3  -1  -3</strong>] 5  3  6  7       -1
- 1  3 [<strong>-1  -3  5</strong>] 3  6  7       -1
- 1  3  -1 [<strong>-3  5  3</strong>] 6  7        3
- 1  3  -1  -3 [<strong>5  3  6</strong>] 7        5
- 1  3  -1  -3  5 [<strong>3  6  7</strong>]       6
-</pre>
-
-<p><strong class="example">Example 2:</strong></p>
-
-<pre>
-<strong>Input:</strong> nums = [1,2,3,4,2,3,1,4,2], k = 3
-<strong>Output:</strong> [2.00000,3.00000,3.00000,3.00000,2.00000,3.00000,2.00000]
-</pre>
-
-<p>&nbsp;</p>
-<p><strong>Constraints:</strong></p>
-
-<ul>
-	<li><code>1 &lt;= k &lt;= nums.length &lt;= 10<sup>5</sup></code></li>
-	<li><code>-2<sup>31</sup> &lt;= nums[i] &lt;= 2<sup>31</sup> - 1</code></li>
-</ul>
-
-<!-- description:end -->
-
-## Solutions
-
-<!-- solution:start -->
-
-### Solution 1: Dual Priority Queues (Min-Heap and Max-Heap) + Lazy Deletion
-
-We can use two priority queues (min-heap and max-heap) to maintain the elements in the current window. One priority queue stores the smaller half of the elements, and the other priority queue stores the larger half of the elements. This way, the median of the current window is either the average of the top elements of the two heaps or one of the top elements.
-
-We design a class $\textit{MedianFinder}$ to maintain the elements in the current window. This class includes the following methods:
-
--   `add_num(num)`: Adds $\textit{num}$ to the current window.
--   `find_median()`: Returns the median of the elements in the current window.
--   `remove_num(num)`: Removes $\textit{num}$ from the current window.
--   `prune(pq)`: If the top element of the heap is in the lazy deletion dictionary $\textit{delayed}$, it pops the top element from the heap and decrements its lazy deletion count. If the lazy deletion count of the element becomes zero, it removes the element from the lazy deletion dictionary.
--   `rebalance()`: If the number of elements in the smaller half exceeds the number of elements in the larger half by $2$, it moves the top element of the larger half to the smaller half. If the number of elements in the smaller half is less than the number of elements in the larger half, it moves the top element of the larger half to the smaller half.
-
-In the `add_num(num)` method, we first consider adding $\textit{num}$ to the smaller half. If $\textit{num}$ is greater than the top element of the larger half, we add $\textit{num}$ to the larger half. Then we call the `rebalance()` method to ensure that the size difference between the two priority queues does not exceed $1$.
-
-In the `remove_num(num)` method, we increment the lazy deletion count of $\textit{num}$. Then we compare $\textit{num}$ with the top element of the smaller half. If $\textit{num}$ is less than or equal to the top element of the smaller half, we update the size of the smaller half and call the `prune()` method to ensure that the top element of the smaller half is not in the lazy deletion dictionary. Otherwise, we update the size of the larger half and call the `prune()` method to ensure that the top element of the larger half is not in the lazy deletion dictionary.
-
-In the `find_median()` method, if the current window size is odd, we return the top element of the smaller half; otherwise, we return the average of the top elements of the smaller half and the larger half.
-
-In the `prune(pq)` method, if the top element of the heap is in the lazy deletion dictionary, it pops the top element from the heap and decrements its lazy deletion count. If the lazy deletion count of the element becomes zero, it removes the element from the lazy deletion dictionary.
-
-In the `rebalance()` method, if the number of elements in the smaller half exceeds the number of elements in the larger half by $2$, it moves the top element of the larger half to the smaller half. If the number of elements in the smaller half is less than the number of elements in the larger half, it moves the top element of the larger half to the smaller half.
-
-The time complexity is $O(n \times \log n)$, and the space complexity is $O(n)$. Here, $n$ is the length of the array $\textit{nums}$.
-
-<!-- tabs:start -->
-
-#### Python3
-
-```python
-class MedianFinder:
-    def __init__(self, k: int):
-        self.k = k
-        self.small = []
-        self.large = []
-        self.delayed = defaultdict(int)
-        self.small_size = 0
-        self.large_size = 0
-
-    def add_num(self, num: int):
-        if not self.small or num <= -self.small[0]:
-            heappush(self.small, -num)
-            self.small_size += 1
-        else:
-            heappush(self.large, num)
-            self.large_size += 1
-        self.rebalance()
-
-    def find_median(self) -> float:
-        return -self.small[0] if self.k & 1 else (-self.small[0] + self.large[0]) / 2
-
-    def remove_num(self, num: int):
-        self.delayed[num] += 1
-        if num <= -self.small[0]:
-            self.small_size -= 1
-            if num == -self.small[0]:
-                self.prune(self.small)
-        else:
-            self.large_size -= 1
-            if num == self.large[0]:
-                self.prune(self.large)
-        self.rebalance()
-
-    def prune(self, pq: List[int]):
-        sign = -1 if pq is self.small else 1
-        while pq and sign * pq[0] in self.delayed:
-            self.delayed[sign * pq[0]] -= 1
-            if self.delayed[sign * pq[0]] == 0:
-                self.delayed.pop(sign * pq[0])
-            heappop(pq)
-
-    def rebalance(self):
-        if self.small_size > self.large_size + 1:
-            heappush(self.large, -heappop(self.small))
-            self.small_size -= 1
-            self.large_size += 1
-            self.prune(self.small)
-        elif self.small_size < self.large_size:
-            heappush(self.small, -heappop(self.large))
-            self.large_size -= 1
-            self.small_size += 1
-            self.prune(self.large)
-
-
-class Solution:
-    def medianSlidingWindow(self, nums: List[int], k: int) -> List[float]:
-        finder = MedianFinder(k)
-        for x in nums[:k]:
-            finder.add_num(x)
-        ans = [finder.find_median()]
-        for i in range(k, len(nums)):
-            finder.add_num(nums[i])
-            finder.remove_num(nums[i - k])
-            ans.append(finder.find_median())
-        return ans
-```
-
-#### Java
-
-```java
-class MedianFinder {
-    private PriorityQueue<Integer> small = new PriorityQueue<>(Comparator.reverseOrder());
-    private PriorityQueue<Integer> large = new PriorityQueue<>();
-    private Map<Integer, Integer> delayed = new HashMap<>();
-    private int smallSize;
-    private int largeSize;
-    private int k;
-
-    public MedianFinder(int k) {
-        this.k = k;
-    }
-
-    public void addNum(int num) {
-        if (small.isEmpty() || num <= small.peek()) {
-            small.offer(num);
-            ++smallSize;
-        } else {
-            large.offer(num);
-            ++largeSize;
-        }
-        rebalance();
-    }
-
-    public double findMedian() {
-        return (k & 1) == 1 ? small.peek() : ((double) small.peek() + large.peek()) / 2;
-    }
-
-    public void removeNum(int num) {
-        delayed.merge(num, 1, Integer::sum);
-        if (num <= small.peek()) {
-            --smallSize;
-            if (num == small.peek()) {
-                prune(small);
-            }
-        } else {
-            --largeSize;
-            if (num == large.peek()) {
-                prune(large);
-            }
-        }
-        rebalance();
-    }
-
-    private void prune(PriorityQueue<Integer> pq) {
-        while (!pq.isEmpty() && delayed.containsKey(pq.peek())) {
-            if (delayed.merge(pq.peek(), -1, Integer::sum) == 0) {
-                delayed.remove(pq.peek());
-            }
-            pq.poll();
-        }
-    }
-
-    private void rebalance() {
-        if (smallSize > largeSize + 1) {
-            large.offer(small.poll());
-            --smallSize;
-            ++largeSize;
-            prune(small);
-        } else if (smallSize < largeSize) {
-            small.offer(large.poll());
-            --largeSize;
-            ++smallSize;
-            prune(large);
-        }
-    }
-}
-
-class Solution {
-    public double[] medianSlidingWindow(int[] nums, int k) {
-        MedianFinder finder = new MedianFinder(k);
-        for (int i = 0; i < k; ++i) {
-            finder.addNum(nums[i]);
-        }
-        int n = nums.length;
-        double[] ans = new double[n - k + 1];
-        ans[0] = finder.findMedian();
-        for (int i = k; i < n; ++i) {
-            finder.addNum(nums[i]);
-            finder.removeNum(nums[i - k]);
-            ans[i - k + 1] = finder.findMedian();
-        }
-        return ans;
-    }
-}
-```
-
-#### C++
-
-```cpp
-class MedianFinder {
-public:
-    MedianFinder(int k) {
-        this->k = k;
-    }
-
-    void addNum(int num) {
-        if (small.empty() || num <= small.top()) {
-            small.push(num);
-            ++smallSize;
-        } else {
-            large.push(num);
-            ++largeSize;
-        }
-        reblance();
-    }
-
-    void removeNum(int num) {
-        ++delayed[num];
-        if (num <= small.top()) {
-            --smallSize;
-            if (num == small.top()) {
-                prune(small);
-            }
-        } else {
-            --largeSize;
-            if (num == large.top()) {
-                prune(large);
-            }
-        }
-        reblance();
-    }
-
-    double findMedian() {
-        return k & 1 ? small.top() : ((double) small.top() + large.top()) / 2.0;
-    }
-
-private:
-    priority_queue<int> small;
-    priority_queue<int, vector<int>, greater<int>> large;
-    unordered_map<int, int> delayed;
-    int smallSize = 0;
-    int largeSize = 0;
-    int k;
-
-    template <typename T>
-    void prune(T& pq) {
-        while (!pq.empty() && delayed[pq.top()]) {
-            if (--delayed[pq.top()] == 0) {
-                delayed.erase(pq.top());
-            }
-            pq.pop();
-        }
-    }
-
-    void reblance() {
-        if (smallSize > largeSize + 1) {
-            large.push(small.top());
-            small.pop();
-            --smallSize;
-            ++largeSize;
-            prune(small);
-        } else if (smallSize < largeSize) {
-            small.push(large.top());
-            large.pop();
-            ++smallSize;
-            --largeSize;
-            prune(large);
-        }
-    }
-};
-
-class Solution {
-public:
-    vector<double> medianSlidingWindow(vector<int>& nums, int k) {
-        MedianFinder finder(k);
-        for (int i = 0; i < k; ++i) {
-            finder.addNum(nums[i]);
-        }
-        vector<double> ans = {finder.findMedian()};
-        for (int i = k; i < nums.size(); ++i) {
-            finder.addNum(nums[i]);
-            finder.removeNum(nums[i - k]);
-            ans.push_back(finder.findMedian());
-        }
-        return ans;
-    }
-};
-```
-
-#### Go
-
-```go
-type MedianFinder struct {
-	small                hp
-	large                hp
-	delayed              map[int]int
-	smallSize, largeSize int
-	k                    int
-}
-
-func Constructor(k int) MedianFinder {
-	return MedianFinder{hp{}, hp{}, map[int]int{}, 0, 0, k}
-}
-
-func (this *MedianFinder) AddNum(num int) {
-	if this.small.Len() == 0 || num <= -this.small.IntSlice[0] {
-		heap.Push(&this.small, -num)
-		this.smallSize++
-	} else {
-		heap.Push(&this.large, num)
-		this.largeSize++
-	}
-	this.rebalance()
-}
-
-func (this *MedianFinder) FindMedian() float64 {
-	if this.k&1 == 1 {
-		return float64(-this.small.IntSlice[0])
-	}
-	return float64(-this.small.IntSlice[0]+this.large.IntSlice[0]) / 2
-}
-
-func (this *MedianFinder) removeNum(num int) {
-	this.delayed[num]++
-	if num <= -this.small.IntSlice[0] {
-		this.smallSize--
-		if num == -this.small.IntSlice[0] {
-			this.prune(&this.small)
-		}
-	} else {
-		this.largeSize--
-		if num == this.large.IntSlice[0] {
-			this.prune(&this.large)
-		}
-	}
-	this.rebalance()
-}
-
-func (this *MedianFinder) prune(pq *hp) {
-	sign := 1
-	if pq == &this.small {
-		sign = -1
-	}
-	for pq.Len() > 0 && this.delayed[sign*pq.IntSlice[0]] > 0 {
-		this.delayed[sign*pq.IntSlice[0]]--
-		if this.delayed[sign*pq.IntSlice[0]] == 0 {
-			delete(this.delayed, sign*pq.IntSlice[0])
-		}
-		heap.Pop(pq)
-	}
-}
-
-func (this *MedianFinder) rebalance() {
-	if this.smallSize > this.largeSize+1 {
-		heap.Push(&this.large, -heap.Pop(&this.small).(int))
-		this.smallSize--
-		this.largeSize++
-		this.prune(&this.small)
-	} else if this.smallSize < this.largeSize {
-		heap.Push(&this.small, -heap.Pop(&this.large).(int))
-		this.smallSize++
-		this.largeSize--
-		this.prune(&this.large)
-	}
-}
-
-func medianSlidingWindow(nums []int, k int) []float64 {
-	finder := Constructor(k)
-	for _, num := range nums[:k] {
-		finder.AddNum(num)
-	}
-	ans := []float64{finder.FindMedian()}
-	for i := k; i < len(nums); i++ {
-		finder.AddNum(nums[i])
-		finder.removeNum(nums[i-k])
-		ans = append(ans, finder.FindMedian())
-	}
-	return ans
-}
-
-type hp struct{ sort.IntSlice }
-
-func (h hp) Less(i, j int) bool { return h.IntSlice[i] < h.IntSlice[j] }
-func (h *hp) Push(v any)        { h.IntSlice = append(h.IntSlice, v.(int)) }
-func (h *hp) Pop() any {
-	a := h.IntSlice
-	v := a[len(a)-1]
-	h.IntSlice = a[:len(a)-1]
-	return v
-}
-```
-
-<!-- tabs:end -->
-
-<!-- solution:end -->
-
-<!-- solution:start -->
-
-### Solution 2: Ordered Set
-
-We can use two ordered sets to maintain the elements in the current window. The ordered set $l$ stores the smaller half of the elements in the current window, and the ordered set $r$ stores the larger half of the elements.
-
-We traverse the array $\textit{nums}$. For each element $x$, we add it to the ordered set $r$, then move the smallest element in the ordered set $r$ to the ordered set $l$. If the size of the ordered set $l$ is greater than the size of the ordered set $r$ by more than $1$, we move the largest element in the ordered set $l$ to the ordered set $r$.
-
-If the total number of elements in the current window is $k$ and the size is odd, the maximum value in the ordered set $l$ is the median. If the size of the current window is even, the average of the maximum value in the ordered set $l$ and the minimum value in the ordered set $r$ is the median. Then, we remove the leftmost element of the window and continue traversing the array.
-
-The time complexity is $O(n \log k)$, and the space complexity is $O(k)$. Here, $n$ is the length of the array $\textit{nums}$.
-
-<!-- tabs:start -->
-
-#### Python3
-
-```python
-from sortedcontainers import SortedList
-
-
-class Solution:
-    def medianSlidingWindow(self, nums: List[int], k: int) -> List[float]:
-        l = SortedList()
-        r = SortedList()
-        ans = []
-        for i, x in enumerate(nums):
-            r.add(x)
-            l.add(r.pop(0))
-            while len(l) - len(r) > 1:
-                r.add(l.pop())
-            j = i - k + 1
-            if j >= 0:
-                ans.append(l[-1] if k & 1 else (l[-1] + r[0]) / 2)
-                if nums[j] in l:
-                    l.remove(nums[j])
-                else:
-                    r.remove(nums[j])
-        return ans
-```
-
-#### Java
-
-```java
-class Solution {
-    public double[] medianSlidingWindow(int[] nums, int k) {
-        TreeMap<Integer, Integer> l = new TreeMap<>();
-        TreeMap<Integer, Integer> r = new TreeMap<>();
-        int n = nums.length;
-        double[] ans = new double[n - k + 1];
-        int lSize = 0, rSize = 0;
-        for (int i = 0; i < n; ++i) {
-            r.merge(nums[i], 1, Integer::sum);
-            int x = r.firstKey();
-            if (r.merge(x, -1, Integer::sum) == 0) {
-                r.remove(x);
-            }
-            l.merge(x, 1, Integer::sum);
-            ++lSize;
-            while (lSize - rSize > 1) {
-                x = l.lastKey();
-                if (l.merge(x, -1, Integer::sum) == 0) {
-                    l.remove(x);
-                }
-                r.merge(x, 1, Integer::sum);
-                --lSize;
-                ++rSize;
-            }
-            int j = i - k + 1;
-            if (j >= 0) {
-                ans[j] = k % 2 == 1 ? l.lastKey() : ((double) l.lastKey() + r.firstKey()) / 2;
-                if (l.containsKey(nums[j])) {
-                    if (l.merge(nums[j], -1, Integer::sum) == 0) {
-                        l.remove(nums[j]);
-                    }
-                    --lSize;
-                } else {
-                    if (r.merge(nums[j], -1, Integer::sum) == 0) {
-                        r.remove(nums[j]);
-                    }
-                    --rSize;
-                }
-            }
-        }
-        return ans;
-    }
-}
-```
-
-#### C++
-
-```cpp
-class Solution {
-public:
-    vector<double> medianSlidingWindow(vector<int>& nums, int k) {
-        multiset<int> l, r;
-        int n = nums.size();
-        vector<double> ans;
-        for (int i = 0; i < n; ++i) {
-            r.insert(nums[i]);
-            l.insert(*r.begin());
-            r.erase(r.begin());
-            while (l.size() - r.size() > 1) {
-                r.insert(*l.rbegin());
-                l.erase(prev(l.end()));
-            }
-            int j = i - k + 1;
-            if (j >= 0) {
-                ans.push_back(k % 2 ? *l.rbegin() : ((double) *l.rbegin() + *r.begin()) / 2);
-                auto it = l.find(nums[j]);
-                if (it != l.end()) {
-                    l.erase(it);
-                } else {
-                    r.erase(r.find(nums[j]));
-                }
-            }
-        }
-        return ans;
-    }
-};
-```
-
-#### Go
-
-```go
-func medianSlidingWindow(nums []int, k int) (ans []float64) {
-	l := redblacktree.New[int, int]()
-	r := redblacktree.New[int, int]()
-	merge := func(st *redblacktree.Tree[int, int], x, v int) {
-		c, _ := st.Get(x)
-		if c+v == 0 {
-			st.Remove(x)
-		} else {
-			st.Put(x, c+v)
-		}
-	}
-	lSize, rSize := 0, 0
-	for i, x := range nums {
-		merge(r, x, 1)
-		x = r.Left().Key
-		merge(r, x, -1)
-		merge(l, x, 1)
-		lSize++
-		for lSize-rSize > 1 {
-			x = l.Right().Key
-			merge(l, x, -1)
-			merge(r, x, 1)
-			lSize--
-			rSize++
-		}
-		if j := i - k + 1; j >= 0 {
-			if k%2 == 1 {
-				ans = append(ans, float64(l.Right().Key))
-			} else {
-				ans = append(ans, float64(l.Right().Key+r.Left().Key)/2)
-			}
-			if x = nums[j]; x <= l.Right().Key {
-				merge(l, x, -1)
-				lSize--
-			} else {
-				merge(r, x, -1)
-				rSize--
-			}
-		}
-	}
-	return
-}
-```
-
-#### TypeScript
-
-```ts
 function medianSlidingWindow(nums: number[], k: number): number[] {
     const l = new TreapMultiSet<number>((a, b) => a - b);
     const r = new TreapMultiSet<number>((a, b) => a - b);
@@ -1265,10 +639,3 @@ class TreapMultiSet<T = number> implements ITreapMultiSet<T> {
         yield* this.reverseInOrder(root.left);
     }
 }
-```
-
-<!-- tabs:end -->
-
-<!-- solution:end -->
-
-<!-- problem:end -->
